requirements.txt
pypdf
langchain-community
python-dotenv
google-generativeai
streamlit
ipykernel
faiss-cpu
